backend/backend/settings_folder/docker_awsdb.py
import os
from pathlib import Path
from decouple import config

# SECURITY WARNING: keep the secret key used in production secret!
SECRET_KEY = config('SECRET_KEY')

# SECURITY WARNING: don't run with debug turned on in production!
SECRET_KEY = config('DEBUG')

ALLOWED_HOSTS = ['*']

CORS_ALLOWED_ORIGINS = [
    'http://localhost',
    'http://127.0.0.1',
    'http://0.0.0.0',
]




# Build paths inside the project like this: BASE_DIR / 'subdir'.
BASE_DIR = Path(__file__).resolve().parent.parent

DATABASES = {
    "default": {
        "ENGINE": "django.db.backends.postgresql",
        "NAME": config('DB_NAME'),
        "USER": config('DB_USER'),
        "PASSWORD": config('DB_PASSWORD'),
        "HOST": config('DB_HOST_LOCAL'),
        "PORT": config('DB_PORT'),
    }
}


STATIC_URL = 'static/'
STATIC_ROOT = '/static/'

MEDIA_URL = 'media/'
MEDIA_ROOT = '/media/'
