backend/requirements.txt
django
django-cors-headers
djangorestframework
gunicorn
python-decouple
pillow
psycopg2-binary